# faster replicate