instead of using memcpy to copy the text n times,
we double the copied text every iteration, thus calling
the super-fast memcpy only ~log n times (with larger and larger chunks to copy)