@kuk0 2017-11-04
before:
```
benchmarking Pure/replicate string/Text+tiny
time                 36.81 ns   (36.79 ns .. 36.84 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 36.37 ns   (36.27 ns .. 36.45 ns)
std dev              279.7 ps   (235.4 ps .. 335.9 ps)

benchmarking Pure/replicate string/Text+ascii
time                 230.1 ms   (229.9 ms .. 230.7 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 231.8 ms   (231.2 ms .. 232.2 ms)
std dev              594.1 μs   (0.0 s .. 685.3 μs)
variance introduced by outliers: 19% (moderately inflated)

benchmarking Pure/replicate string/Text+russian
time                 49.20 μs   (45.03 μs .. 53.98 μs)
                     0.959 R²   (0.945 R² .. 0.999 R²)
mean                 47.25 μs   (45.60 μs .. 50.01 μs)
std dev              7.410 μs   (4.832 μs .. 9.980 μs)
variance introduced by outliers: 93% (severely inflated)

benchmarking Pure/replicate string/Text+japanese
time                 49.93 μs   (48.25 μs .. 53.28 μs)
                     0.980 R²   (0.957 R² .. 0.999 R²)
mean                 49.88 μs   (48.45 μs .. 52.90 μs)
std dev              6.588 μs   (3.935 μs .. 9.819 μs)
variance introduced by outliers: 90% (severely inflated)
```

after:
```
benchmarking Pure/replicate string/Text+tiny
time                 36.31 ns   (36.14 ns .. 36.66 ns)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 27.30 ns   (26.47 ns .. 28.00 ns)
std dev              1.176 ns   (841.7 ps .. 1.586 ns)
variance introduced by outliers: 64% (severely inflated)

benchmarking Pure/replicate string/Text+ascii
time                 33.28 μs   (33.18 μs .. 33.48 μs)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 25.94 μs   (25.18 μs .. 26.57 μs)
std dev              1.330 μs   (965.3 ns .. 1.945 μs)
variance introduced by outliers: 54% (severely inflated)

benchmarking Pure/replicate string/Text+russian
time                 4.656 μs   (4.638 μs .. 4.673 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 3.810 μs   (3.680 μs .. 3.917 μs)
std dev              295.6 ns   (242.4 ns .. 363.1 ns)
variance introduced by outliers: 79% (severely inflated)

benchmarking Pure/replicate string/Text+japanese
time                 4.983 μs   (4.969 μs .. 5.000 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 4.123 μs   (3.993 μs .. 4.237 μs)
std dev              305.8 ns   (254.9 ns .. 370.3 ns)
variance introduced by outliers: 78% (severely inflated)
```